+++ R/inframe.R
@@ -73,7 +73,7 @@ sim_inframe <- function(df_visit, r = 1000, df_site = NULL, event_names = c("ae"
   # aggregate per patient to get max visits
   df_pat_aggr_pool <- pat_aggr(df_visit)
 
-  colnames_event <- paste0(event_names, "_events")
+  colnames_event <- paste0(event_names, "_count")
   colnames_low <- paste0(event_names, "_prob_low")
   colnames_ori <- paste0(event_names, "_per_visit_site_ori")
   colnames_rep <- paste0(event_names, "_per_visit_rep")