+++ R/zzz_test_helper.R
@@ -11,7 +11,8 @@ get_df_visit_test_mapped <- function() {
     visit = "visit"
   )
 
-  df_visit <- map_col_names(df_visit, col_names)
+  df_visit <- map_col_names(df_visit, col_names) %>%
+    rename(n_ae = n_event)
 
 }
 
@@ -25,7 +26,7 @@ get_df_visit_test <- function() {
     n_pat = 100,
     n_sites = 5,
     ratio_out = 0.4,
-    factor_event_rate = 0.6
+    factor_event_rate = - 0.4 # 40% under-reporting
   )
 
   df_visit1$study_id <- "A"
@@ -36,7 +37,7 @@ get_df_visit_test <- function() {
     n_pat = 100,
     n_sites = 5,
     ratio_out = 0.2,
-    factor_event_rate = 0.1
+    factor_event_rate = - 0.1 # 10% under-reporting
   )
 
   df_visit2$study_id <- "B"
@@ -46,38 +47,6 @@ get_df_visit_test <- function() {
   return(df_visit_test)
 }
 
-#' Get df_portf_test
-#' @keywords internal
-get_df_portf_test <- function() {
-
-  col_names <- c(
-    study_id = "study_id",
-    site_id = "site_id",
-    patient_id = "patient_id",
-    visit = "visit",
-    n_ae = "n_event"
-  )
-
-  df_site_max_test <- get_df_visit_test() %>%
-    map_col_names(col_names) %>%
-    group_by(.data$study_id, .data$site_number, .data$patnum) %>%
-    summarise(max_visit = max(.data$visit),
-              max_ae = max(.data$n_ae),
-              .groups = "drop")
-
-  df_config_test <- simaerep::get_config(
-    df_site_max_test,
-    anonymize = TRUE,
-    min_pat_per_study = 100,
-    min_sites_per_study = 5
-  )
-
-  set.seed(1)
-
-  df_portf_test <- sim_test_data_portfolio(df_config_test, progress = FALSE)
-
-
-}
 
 #' @title simulate test data events
 #' @description generates multi-event data using sim_test_data_study()
@@ -98,13 +67,15 @@ sim_test_data_events <- function(
 
   set.seed(1)
   df_visit3 <- sim_test_data_study(n_pat = 100, n_sites = n_sites,
-                                   frac_site_with_ur = 0.4, ur_rate = 0.6,
-                                   event_per_visit_mean = event_per_visit_mean, event_names = event_names)
+                                   ratio_out = 0.4, factor_event_rate = - 0.6,
+                                   event_per_visit_mean = event_per_visit_mean,
+                                   event_names = event_names)
   df_visit3$study_id <- "A"
   set.seed(2)
   df_visit4 <- sim_test_data_study(n_pat = 100, n_sites = n_sites,
-                                   frac_site_with_ur = 0.2, ur_rate = 0.1,
-                                   event_per_visit_mean = event_per_visit_mean, event_names = event_names)
+                                   ratio_out = 0.2, factor_event_rate = - 0.1,
+                                   event_per_visit_mean = event_per_visit_mean,
+                                   event_names = event_names)
   df_visit4$study_id <- "B"
   df_visit_events_test <- dplyr::bind_rows(df_visit3, df_visit4)
   return(df_visit_events_test)