+++ tests/testthat/test_S3_orivisit.R
@@ -70,3 +70,4 @@ test_that("as.data.frame.orivisit() stops when str_call is not in the parent env
   )
   expect_error(as.data.frame.orivisit(x), regex = message)
 })
+