+++ tests/testthat/test_check_df_visit.R
@@ -6,7 +6,7 @@ test_that("check_df_visit() must fix missing visits and throw a warning", {
     filter(visit != 3)
 
   # nolint start
-  expect_warning({df_visit_corr <- check_df_visit(df_visit_filt)})
+  expect_warning({df_visit_corr <- check_df_visit(df_visit_filt, event_names = "ae")})
   expect_true(3 %in% df_visit_corr$visit)
   expect_true(nrow(df_visit_corr) == nrow(df_visit))
   # nolint end
@@ -16,7 +16,7 @@ test_that("check_df_visit() must fix missing visits and throw a warning", {
 test_that("check_df_visit() must aggregate duplicated visits and throw a warning", {
 
   # nolint start
-  expect_warning({df_visit_corr <- check_df_visit(bind_rows(df_visit, df_visit))})
+  expect_warning({df_visit_corr <- check_df_visit(bind_rows(df_visit, df_visit), event_names = "ae")})
   expect_true(nrow(df_visit_corr) == nrow(df_visit))
   # nolint end
 
@@ -25,8 +25,8 @@ test_that("check_df_visit() must aggregate duplicated visits and throw a warning
 test_that("check_df_visit() must thrown an error when n_ae and visit columns are not numeric", {
   expect_error({
     df_visit %>%
-      mutate_at(vars(n_event, visit), as.character) %>%
-      check_df_visit()
+      mutate_at(vars(n_ae, visit), as.character) %>%
+      check_df_visit(event_names = "ae")
   },
   regexp = "event number and visit columns must be numeric"
   )
@@ -37,9 +37,9 @@ test_that("check_df_visit() must thrown an error when NA's are detected", {
     df_visit %>%
       summarise_all(~ NA) %>%
       bind_rows(df_visit) %>%
-      check_df_visit()
+      check_df_visit(event_names = "ae")
   },
-  regexp = "NA detected in columns: study_id, site_number, patnum, n_event, visit"
+  regexp = "NA detected in columns: study_id, site_number, patnum, n_ae, visit"
   )
 })