+++ NEWS.md
@@ -2,8 +2,14 @@
 - use get_cum_mean_event_dev() instead of get_site_mean_ae_dev() for plotting which does not rely on visit_med75 to reduce noise
 - eval_sites will combine over- and under-reporting probability into one column prob
 - eval_sites with inframe method will calculate delta events
-- new simaerep defaults inframe = TRUE, mult_corr = FALSE, visit_med75 = FALSE, under_only = FALSE
-- retire sim_ur_scenarios and get_portf_perf, portfolio performance will be calculated by removing events directly from the data with sim_ur
+- new simaerep defaults inframe = TRUE, mult_corr = TRUE, visit_med75 = FALSE, under_only = FALSE
+- retire sim_ur_scenarios, get_portf_perf, sim_studies, get_ecd_values, get_pat_pool_config portfolio performance will be calculated by removing events directly from the data with sim_ur
+- remove check argument from low-level interface functions site_aggr and sim_sites
+- use test helper functions to generate data for unit testing
+- reuse saved test data for regression testing towards earlier simaerep versions
+- pass alternative column names in df_visit input data to simaerep via col_names argument and preserve names in simaerep results
+- improved print output for simaerep and orivist S3 objects
+- adapt validation unit tests to include event over and under reporting
 
 # simaerep 0.7.0
 - event_names argument, allow to calculate multiple reporting probabilities at once https://github.com/openpharma/simaerep/issues/73